Move MainActivity to com.ncnf package
