+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:theme="@style/Theme.Bootcamp"
         android:usesCleartextTraffic="true">
 
-        <activity android:name="com.ncnf.main.MainActivity" >
+        <activity android:name="com.ncnf.MainActivity" >
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -32,15 +32,15 @@
 
         <activity
             android:name="com.ncnf.user.UserProfileActivity"
-            android:parentActivityName="com.ncnf.main.MainActivity"
+            android:parentActivityName="com.ncnf.MainActivity"
             tools:ignore="MissingClass" />
         <activity
             android:name="com.ncnf.authentication.SignUpActivity"
-            android:parentActivityName="com.ncnf.main.MainActivity"
+            android:parentActivityName="com.ncnf.MainActivity"
             tools:ignore="MissingClass" />
         <activity
             android:name="com.ncnf.authentication.SignInActivity"
-            android:parentActivityName="com.ncnf.main.MainActivity"
+            android:parentActivityName="com.ncnf.MainActivity"
             tools:ignore="MissingClass" />
 
         <!--
@@ -58,7 +58,7 @@
         <activity
             android:name="com.ncnf.map.MapActivity"
             android:label="@string/title_activity_maps"
-            android:parentActivityName="com.ncnf.main.MainActivity" />
+            android:parentActivityName="com.ncnf.MainActivity" />
 
         <uses-library
             android:name="com.google.android.wearable"
@@ -75,7 +75,7 @@
         <activity
             android:name="com.ncnf.feed.FeedActivity"
             android:label="@string/title_activity_feed"
-            android:parentActivityName="com.ncnf.main.MainActivity" >
+            android:parentActivityName="com.ncnf.MainActivity" >
         </activity>
 
         <activity
